pyproject.toml
[tool.poetry]
name = "onnx2kerastl"
version = "0.0.143"
description = ""
authors = ["dorhar <doron.harnoy@tensorleap.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
tensorflow = {version = "2.11.0", markers = "platform_machine  == 'x86_64'"}
tensorflow-macos = {version = "2.11.0", markers = "platform_machine  == 'arm64'"}
onnx = "1.10.1"
protobuf = "^3.19.6"
tensorflow-addons = {version = "^0.19.0", markers = "platform_machine  == 'x86_64'"}
numpy = "1.23.5"
fvcore = "^0.1.5.post20221221"
boto3 = "^1.24.22"
tensorflow-io-gcs-filesystem = "0.30.0"
keras-data-format-converter = "0.1.17"
tf = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
torch = {version = "1.12.1", markers = "platform_machine  == 'x86_64'"}
torchvision = {version = "^0.12.0", markers = "platform_machine  == 'x86_64'"}
transformers = {extras = ["onnx"], version = "^4.25.1", markers = "platform_machine  == 'x86_64'"}
sentencepiece ={version = "^0.1.96", markers = "platform_machine  == 'x86_64'"}
pandas = "^2.0.3"
datasets = "^2.14.1"
librosa = "^0.10.0.post2"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
